refactored endpoints
