+++ src_ts/components/pages/interventions/intervention-tab-pages/utils/fire-custom-event.ts
@@ -1,4 +1,4 @@
-import {AnyObject} from '../../types/globals';
+import {AnyObject} from './types';
 
 export const fireEvent = (el: HTMLElement, eventName: string, eventDetail?: AnyObject) => {
   el.dispatchEvent(
